html/semantics/edits/attributes-common-to-ins-and-del-elements/HTMLModElement.html

<!DOCTYPE html>
<meta charset=UTF-8>
<title>HTML Test: HTMLModElement</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/edits.html#attributes-common-to-ins-and-del-elements">
<link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/infrastructure.html#valid-url-potentially-surrounded-by-spaces">
<link rel="help" href="http://www.w3.org/TR/2012/CR-html5-20121217/infrastructure.html#valid-date-string-with-optional-time">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>
  var runTest = function(testElement, ele) {
    // elements have been tested by ../../interfaces.html

    // cite
    test(function() {
      assert_true("cite" in testElement, "IDL attribute exists");
      assert_equals(typeof testElement["cite"], "string", "typeof IDL attribute");
      assert_equals(testElement["cite"], "", "IDL get with DOM attribute unset");
    }, ele + ".cite IDL attribute checking for existence, type and default value");

    [ // [valid URL potentially surrounded by spaces, expected resolved URL]
      // TODO: add more valid URL in need; and make better URL assertion
      ["http://www.w3.org/TR/2012/CR-html5-20121217/", "http://www.w3.org/TR/2012/CR-html5-20121217/"],
      [" http://www.w3.org/TR/2012/CR-html5-20121217/ ", "http://www.w3.org/TR/2012/CR-html5-20121217/"],
    ].forEach(function(data) {
      var cite = data[0];
      var expect = data[1];
      
      testElement.setAttribute("cite", cite);
      test(function() {
        assert_equals(testElement["cite"], expect, "IDL get");
      }, ele + ".cite setAttribute() to " + cite + " followed by IDL get");
      test(function() {
        assert_equals(testElement.getAttribute("cite"), cite, "getAttribute()");
      }, ele + ".cite setAttribute() to " + cite + " followed by getAttribute()");

      test(function() {
        try {
          testElement["cite"] = cite;
          test(function() {
            assert_equals(testElement["cite"], expect, "IDL get");
          }, ele + ".cite IDL set to " + cite + " followed by IDL get");
          test(function() {
            assert_equals(testElement.getAttribute("cite"), cite, "getAttribute()");
          }, ele + ".cite IDL set to " + cite + " followed by getAttribute()");
        } catch(e) {
          assert_unreached("Should not throw exception: " + e.message);
        }
      }, ele + ".cite IDL set to " + cite + " should not throw exception");
    });

    // dateTime
    test(function() {
      var testElement = document.createElement("ins");
      assert_true("dateTime" in testElement, "IDL attribute exists");
      assert_equals(typeof testElement["dateTime"], "string", "typeof IDL attribute");
      assert_equals(testElement["dateTime"], "", "IDL get with DOM attribute unset");
    }, ele + ".dateTime IDL attribute checking for existence, type and default value");

    [ // valid date string with optional time
      // valid date string
      "2013-07-13",
      // valid global date and time string
      "0037-12-13 00:00Z",
      "2008-02-14T12:02Z",
      "1979-10-14T12:00:00.001-04:00",
      "8592-01-01T02:09+02:09"
    ].forEach(function(dateTime) {
      testElement.setAttribute("dateTime", dateTime);
      test(function() {
        assert_equals(testElement["dateTime"], dateTime, "IDL get");
      }, ele + ".dateTime setAttribute() to " + dateTime + " followed by IDL get");
      test(function() {
        assert_equals(testElement.getAttribute("dateTime"), dateTime, "getAttribute()");
      }, ele + ".dateTime setAttribute() to " + dateTime + " followed by getAttribute()");

      test(function() {
        try {
          testElement["dateTime"] = dateTime;
          test(function() {
            assert_equals(testElement["dateTime"], dateTime, "IDL get");
          }, ele + ".dateTime IDL set to " + dateTime + " followed by IDL get");
          test(function() {
            assert_equals(testElement.getAttribute("dateTime"), dateTime, "getAttribute()");
          }, ele + ".dateTime IDL set to " + dateTime + " followed by getAttribute()");
        } catch(e) {
          assert_unreached("Should not throw exception: " + e.message);
        }
      }, ele + ".dateTime IDL set to " + dateTime + " should not throw exception");
    });
  };
      
  runTest(document.createElement("ins"), "ins");
  runTest(document.createElement("del"), "del");
</script>
